backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "trandat",email = "giavien2002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "python-multipart (>=0.0.24,<0.0.25)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "bcrypt (==4.0.1)"
]

[tool.poetry]
packages = [{include = "backend", from = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
